[REFACTOR/#111] market repository 분리
