+++ app/src/main/java/com/napzak/market/data/product/di/RepositoryModule.kt
@@ -1,7 +1,9 @@
 package com.napzak.market.data.product.di
 
 import com.napzak.market.data.product.repositoryimpl.ExploreProductRepositoryImpl
+import com.napzak.market.data.product.repositoryimpl.MarketProductRepositoryImpl
 import com.napzak.market.domain.explore.repository.ExploreProductRepository
+import com.napzak.market.domain.explore.repository.MarketProductRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -17,4 +19,10 @@ abstract class RepositoryModule {
     abstract fun bindExploreRepository(
         exploreProductRepositoryImpl: ExploreProductRepositoryImpl
     ): ExploreProductRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindMarketProductRepository(
+        marketProductRepositoryImpl: MarketProductRepositoryImpl
+    ): MarketProductRepository
 }
