+++ app/src/main/java/com/napzak/market/domain/explore/repository/MarketProductRepository.kt
@@ -1,0 +1,10 @@
+package com.napzak.market.domain.explore.repository
+
+import com.napzak.market.domain.marketinfo.model.ProductItem
+import com.napzak.market.domain.marketinfo.model.ProductListFilter
+
+interface MarketProductRepository {
+    suspend fun fetchMarketProductSellItems(request: ProductListFilter): Result<List<ProductItem>>
+
+    suspend fun fetchMarketProductBuyItems(request: ProductListFilter): Result<List<ProductItem>>
+}
