+++ app/src/main/java/com/napzak/market/domain/marketinfo/repository/MarketInfoRepository.kt
@@ -6,8 +6,4 @@ import com.napzak.market.domain.marketinfo.model.ProductListFilter
 
 interface MarketInfoRepository {
     suspend fun fetchMarketInfo(request: Long): Result<MarketInfo>
-
-    suspend fun fetchMarketProductSellItems(request: ProductListFilter): Result<List<ProductItem>>
-
-    suspend fun fetchMarketProductBuyItems(request: ProductListFilter): Result<List<ProductItem>>
 }
